Fix issue with interactive installation on Apache
Remove an odd check on the Apache configuration file requiring it to have a
LoadModule directive. The file may not have it and the is no reason to require
it. A clean installation of Apache2.4 on Ubuntu 15.10 through apt-get produces a
configuration file without the directive, causing the failure of the WPA
installation.